Create Calculator component
